backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Werkzeug==2.3.7
authlib==1.3.0
PyJWT==2.8.0
requests==2.31.0
